fix dev env for token, restructure config folder
