+++ src/main/java/com/codurance/retropolis/config/database/DatabaseConfig.java
@@ -1,5 +1,6 @@
-package com.codurance.retropolis.config;
+package com.codurance.retropolis.config.database;
 
+import com.codurance.retropolis.config.Environment;
 import javax.sql.DataSource;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
