+++ src/main/java/com/codurance/retropolis/config/web/WebConfig.java
@@ -1,13 +1,12 @@
-package com.codurance.retropolis.config;
+package com.codurance.retropolis.config.web;
 
-import java.util.Collections;
+import static java.util.Collections.singletonList;
+
+import com.codurance.retropolis.config.Environment;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.core.Ordered;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -28,16 +27,8 @@ public class WebConfig {
 
     @Bean
     public FilterRegistrationBean<CorsFilter> simpleCorsFilter() {
-      UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-      CorsConfiguration config = new CorsConfiguration();
-      config.setAllowCredentials(true);
-      config.setAllowedOrigins(Collections.singletonList("http://retropolis-fe.s3-website.eu-west-2.amazonaws.com"));
-      config.setAllowedMethods(Collections.singletonList("*"));
-      config.setAllowedHeaders(Collections.singletonList("*"));
-      source.registerCorsConfiguration("/**", config);
-      FilterRegistrationBean<CorsFilter> bean = new FilterRegistrationBean<>(new CorsFilter(source));
-      bean.setOrder(Ordered.HIGHEST_PRECEDENCE);
-      return bean;
+      return CorsFilterConfiguration.simpleCorsFilter(singletonList("http://retropolis-fe.s3-website.eu-west-2.amazonaws.com"));
     }
+
   }
 }
