+++ src/main/java/com/codurance/retropolis/config/web/WebDevConfig.java
@@ -1,9 +1,13 @@
-package com.codurance.retropolis.config;
+package com.codurance.retropolis.config.web;
 
+import static java.util.Collections.singletonList;
+
+import com.codurance.retropolis.config.Environment;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.filter.CorsFilter;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import springfox.documentation.builders.PathSelectors;
@@ -29,11 +33,9 @@ public class WebDevConfig {
   @Configuration
   public class WebConfiguration implements WebMvcConfigurer {
 
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-      registry.addMapping("/**")
-          .allowedMethods("*")
-          .allowedOrigins("http://localhost:3000");
+    @Bean
+    public FilterRegistrationBean<CorsFilter> simpleCorsFilter() {
+      return CorsFilterConfiguration.simpleCorsFilter(singletonList("http://localhost:3000"));
     }
 
     @Override
