+++ src/main/java/com/codurance/retropolis/controllers/BoardController.java
@@ -1,7 +1,7 @@
 package com.codurance.retropolis.controllers;
 
 
-import com.codurance.retropolis.config.GoogleTokenAuthenticator;
+import com.codurance.retropolis.config.web.GoogleTokenAuthenticator;
 import com.codurance.retropolis.entities.Board;
 import com.codurance.retropolis.requests.NewBoardRequestObject;
 import com.codurance.retropolis.services.BoardService;
